Moved auth out of desktop and other stuff around
