+++ src/auth/ResetPasswordModal.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const ResetPasswordForm = () => {};
+
+const EnterResetCodeForm = () => {};
+
+const EnterNewPasswordForm = () => {};
+
+const ResetPasswordModal = () => {};
+
+export default ResetPasswordModal;
