+++ src/desktop/components/testPage/TestPage.jsx
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 
 /* Child Components */
-import Textbox from '../common/Textbox';
+import Textbox from '../../../basicComponents/Textbox';
 import TabContainer from '../common/TabContainer';
 import Button from '../../../basicComponents/Button';
 import ProgressBar from '../../../basicComponents/ProgressBar';
